+++ apis/metal3.io/v1alpha1/profile/profile.go
@@ -1,4 +1,4 @@
-package hardware
+package profile
 
 import (
 	"fmt"
